Move DependencyChecker to dependenciesModule
